+++ pyobs/modules/acquisition/base.py
@@ -5,10 +5,12 @@ from astropy.coordinates import SkyCoord, AltAz
 from astropy.wcs import WCS
 import astropy.units as u
 
-from pyobs.interfaces import ITelescope, ICamera, IAcquisition, IRaDecOffsets, IAltAzOffsets, ICameraExposureTime
+from pyobs.interfaces import ITelescope, ICamera, IAcquisition, IRaDecOffsets, IAltAzOffsets, ICameraExposureTime, \
+    IImageType
 from pyobs import Module
 from pyobs.mixins import CameraSettingsMixin
 from pyobs.modules import timeout
+from pyobs.utils.enums import ImageType
 from pyobs.utils.images import Image
 from pyobs.utils.publisher import CsvPublisher
 from pyobs.utils.time import Time
@@ -92,13 +94,15 @@ class BaseAcquisition(Module, CameraSettingsMixin, IAcquisition):
 
         # try given number of attempts
         for a in range(self._attempts):
-            # set exposure time and take image
+            # set exposure time and image type and take image
             if isinstance(camera, ICameraExposureTime):
                 log.info('Exposing image for %.1f seconds...', exposure_time)
                 camera.set_exposure_time(exposure_time).wait()
             else:
                 log.info('Exposing image...')
-            filename = camera.expose(ICamera.ImageType.ACQUISITION).wait()
+            if isinstance(camera, IImageType):
+                camera.set_image_type(ImageType.ACQUISITION)
+            filename = camera.expose().wait()
 
             # download image
             log.info('Downloading image...')