+++ pyobs/modules/focus/focusseries.py
@@ -4,11 +4,12 @@ import threading
 import numpy as np
 
 from pyobs.comm import RemoteException
-from pyobs.interfaces import IFocuser, ICamera, IAutoFocus, IFilters
+from pyobs.interfaces import IFocuser, ICamera, IAutoFocus, IFilters, ICameraExposureTime, IImageType
 from pyobs.events import FocusFoundEvent
 from pyobs import Module, get_object
 from pyobs.mixins import CameraSettingsMixin
 from pyobs.modules import timeout
+from pyobs.utils.enums import ImageType
 from pyobs.utils.focusseries import FocusSeries
 
 log = logging.getLogger(__name__)
@@ -141,7 +142,11 @@ class AutoFocusSeries(Module, CameraSettingsMixin, IAutoFocus):
             if self._abort.is_set():
                 raise InterruptedError()
             try:
-                filename = camera.expose(exposure_time=exposure_time, image_type=ICamera.ImageType.FOCUS).wait()
+                if isinstance(camera, ICameraExposureTime):
+                    camera.set_exposure_time(exposure_time)
+                if isinstance(camera, IImageType):
+                    camera.set_image_type(ImageType.FOCUS)
+                filename = camera.expose().wait()
             except RemoteException:
                 log.error('Could not take image.')
                 continue