+++ tests/modules/camera/test_basecamera.py
@@ -8,6 +8,7 @@ from pyobs.comm.dummy import DummyComm
 from pyobs.environment import Environment
 from pyobs.interfaces import ICamera
 from pyobs.modules.camera import BaseCamera
+from pyobs.utils.enums import ImageType
 
 
 def test_open_close():
@@ -139,7 +140,9 @@ def test_expose():
     assert ICamera.ExposureStatus.IDLE == camera.get_exposure_status()
 
     # expose
-    camera.expose(exposure_time=0, image_type=ICamera.ImageType.OBJECT)
+    camera.set_exposure_time(0)
+    camera.set_image_type(ImageType.OBJECT)
+    camera.expose()
 
     # status must be idle again
     assert ICamera.ExposureStatus.IDLE == camera.get_exposure_status()
@@ -163,7 +166,8 @@ def test_abort():
     def expose():
         with pytest.raises(ValueError):
             camera.set_exposure_time(1.)
-            camera.expose(image_type='object')
+            camera.set_image_type(ImageType.OBJECT)
+            camera.expose()
 
     # expose
     thread = threading.Thread(target=expose)